One more refactor - move exception handlers to distinct file
